requirements.txt
Django==2.2.5
djangorestframework==3.10.3
gunicorn==20.0.0
pytz==2020.1
sqlparse==0.3.1
